+++ src/utils/getDate.ts
@@ -1,4 +1,4 @@
-import {format, getHours, getMinutes} from 'date-fns';
+import {endOfWeek, format, getHours, getMinutes, startOfWeek} from 'date-fns';
 import {DateParsable} from 'react-native-calendar-picker';
 
 const getDateRange = (startDate: DateParsable, endDate: DateParsable) => {
@@ -18,4 +18,11 @@ const getTime = (date: Date) => {
   return '';
 };
 
-export {getDateRange, getTime};
+const getVoteDateRange = () => {
+  const startDate = format(startOfWeek(new Date()), 'M월 dd일');
+  const endDate = format(endOfWeek(new Date()), 'M월 dd일');
+
+  return {startDate, endDate};
+};
+
+export {getDateRange, getTime, getVoteDateRange};
